Retain error return value in hx711_sample_fetch
Before this commit, if any error has happend in the function
hx711_sample_fetch before the exit label, the return error would be
replaced with the return value from the function
gpio_pin_interrupt_configure. This causes returning success value
instead of the error that happend which is misleading.

This commit adds another return error var for the interrupt configure
function so we can still retain the previous return value.

Signed-off-by: Moranr296 <moriki@gmail.com>